tests/wpt/tests/css/css-flexbox/getcomputedstyle/flexbox_computedstyle_flex-basis-percent.html

<!DOCTYPE html>
<title>flexbox | computed style | flex-basis: percent</title>
<link rel="author" href="http://opera.com" title="Opera Software">
<link rel="help" href="http://www.w3.org/TR/css-flexbox-1/#flex-basis-property">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name="flags" content="dom">
<style>
html {
	display: flex;
	width: 100px;
}
body {
	color: red;
	margin: 0;
	flex-basis: 100%;
}
</style>
<div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-basis"),
	          "100%");
});
</script>
